初始化asset服务

+++ admin-api/internal/types/types.go
@@ -1803,7 +1803,7 @@ type UpdateUserBaseResp struct {
 type UpdateUserLevelReq struct {
 	TenantId    int64 `json:"tenantId"`
 	UserId      int64 `path:"userId"`
-	MemberLevel int32 `json:"memberLevel"`
+	MemberLevel int64 `json:"memberLevel"`
 }
 
 type UpdateUserStatusReq struct {
@@ -1974,8 +1974,8 @@ type UserSecurity struct {
 	UserId          int64 `json:"userId"`
 	HasPayPassword  bool  `json:"hasPayPassword"`
 	GoogleEnabled   bool  `json:"googleEnabled"`
-	LoginErrorCount int32 `json:"loginErrorCount"`
-	PayErrorCount   int32 `json:"payErrorCount"`
+	LoginErrorCount int64 `json:"loginErrorCount"`
+	PayErrorCount   int64 `json:"payErrorCount"`
 	LockUntil       int64 `json:"lockUntil,optional"`
 	RiskLevel       int64 `json:"riskLevel"`
 	CreateTimes     int64 `json:"createTimes,optional"`
